Add Forwarding part to Full.js test.
Although the test is not working because of the {from: Forwarding.address} clause raising an 'Unrecognized account' error, the forwarding process that should be in the test was added.
This error makes no sense because the DAO requirement for setPricePerMB was removed and the Forwarding account is the same as the ReserveAccount, maybe it has to do with Truffle test
contracts' pre-deployment before each test.